requirements.txt
cv2
requests
numpy
pillow
opencv_python
